\name{FrF2.catlg128-package}
\alias{FrF2.catlg128-package}
\alias{FrF2.catlg128}
\docType{package}
\title{
Complete catalogues of resolution IV 128 run 2-level fractional factorials up to 24 factors
}
\description{
Description: This package provides the complete catalogues of resolution IV 
regular fractional factorial designs in 128 runs for up to 24 2-level factors.}
\details{
The complete set of designs provided here has been provided by Xu (2009) 
and is useful for special tasks to 
be accomplished with package \pkg{\link[FrF2:FrF2-package]{FrF2}}. The catalogues 
are separately provided here, because they are very large 
and should not be forced upon \pkg{FrF2} users 
who do not need them. 

}
\author{
Ulrike Groemping

Maintainer: Ulrike Groemping <groemping@bht-berlin.de>
}
\references{
   Xu, H. (2009) Algorithmic Construction of Efficient Fractional Factorial Designs 
         With Large Run Sizes. \emph{Technometrics} \bold{51}, 262-277.
}
\keyword{ array }
\keyword{ design }
\seealso{
\code{\link{Catalogues}}
}
